Refactor: Styles 폴더 삭제 및 Components 폴더 정리
