+++ src/components/Footer.js
@@ -2,7 +2,7 @@ import instagram from '../images/instagram.svg';
 import facebook from '../images/facebook.svg';
 import twitter from '../images/twitter.svg';
 import youtube from '../images/youtube.svg';
-import '../styles/Footer.css';
+import './Footer.css';
 
 export default function Footer() {
   return (
