+++ src/components/common/Button.js
@@ -1,4 +1,4 @@
-import '../styles/Button.css';
+import './Button.css';
 
 export default function Button({ content }) {
   return <button>{content}</button>;
